ref(movie page): change hash handler and move out reviews stub
