+++ project/src/components/movie-page-details/movie-page-details.tsx
