+++ project/src/components/tabs/current-tab.tsx
@@ -1,8 +1,8 @@
 import { Film } from '../../types/film';
 import { MoviePageTabNames } from '../../types/movie-page-tab-names';
+import MoviePageDetails from '../movie-page-details/movie-page-details';
 import MoviePageOverview from '../movie-page-overview/movie-page-overview';
 import MovieReviewPage from '../movie-review-page/movie-review-page';
-import MoviePageDetails from '../movie=page-details/movie-page-details';
 
 type currentTabProps = {
     activeTab: string,
@@ -10,8 +10,6 @@ type currentTabProps = {
 };
 
 function CurrentTab({activeTab, film}: currentTabProps): JSX.Element {
-  // eslint-disable-next-line no-console
-  console.log(film);
   switch (activeTab) {
     case MoviePageTabNames.Details:
       return <MoviePageDetails film={film} />;
